Implemented data classes, implemented about half of build index function.
Made a class to hold an encrypted document and a class to hold a secure index.
Implemented part of the build index function for one document up to the point
when all the codewords for that document are to be inserted in the bloom filter.
Very crude, little to no error checking, print debug statements throughout and
rambling comments abound.