Refactored into create_trapdoor and create_codeword
Refactored the body of build index into these two main
steps.  Now I can change search index so there is a version
in this class and one in the client class.  And I can move
the build function to the client class and move part of the
search index to the client class.  This should not be in this
class.