Made Several Modifications
Changed CollectionCreator.parse_directory() by swapping the order
of calls to build an index and to add to a collection.  A document's
index must be built after it is added to a collection.

Fixed code so that each Doc gets a unique id.  Previously, all docs
in all collections had the id of 1.

Changed CollectionCreator.calc_bf_size() to consider how many master
keys are used when deciding how large to make the bloom filter.  This
is because for 9 keys, 9 items are inserted into the bloom filter for
every word.